# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for managing administrative resources like users, servers, employees, groups, etc.",
        "title": "Admin API Documentation",
        "termsOfService": "http://yourdomain.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://yourdomain.com/support",
            "email": "support@yourdomain.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/login": {
            "put": {
                "description": "Accepts JWT using PUT. Returns a new JWT if valid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Renew token via PUT",
                "responses": {
                    "200": {
                        "description": "JWT token returned",
                        "schema": {
                            "$ref": "#/definitions/models.JWTResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid fields",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid credentials)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Accepts JSON credentials (username + password) and returns a token if valid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate a user and issue a JWT (POST)",
                "parameters": [
                    {
                        "description": "Username and password",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token returned",
                        "schema": {
                            "$ref": "#/definitions/models.JWTResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid fields",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid credentials)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Accepts JSON input with username and password to create a new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user account",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully registered",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Invalid input JSON or missing fields",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error contains a descriptive error message.",
                    "type": "string"
                }
            }
        },
        "models.JWTResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Token is the JWT token assigned to the authenticated user.",
                    "type": "string"
                }
            }
        },
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password is the user's password used for authentication.",
                    "type": "string"
                },
                "totp_code": {
                    "description": "TotpCode is required when Time-based OTP is enabled for the account.",
                    "type": "string"
                },
                "username": {
                    "description": "Username is the unique identifier for the user attempting to log in.",
                    "type": "string"
                }
            }
        },
        "models.Role": {
            "type": "string",
            "enum": [
                "user",
                "reviewer",
                "admin",
                "*"
            ],
            "x-enum-varnames": [
                "UserRole",
                "ReviewerRole",
                "AdminRole",
                "DefaultRoleWhitelist"
            ]
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Audit",
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "edited_by": {
                    "type": "string"
                },
                "email": {
                    "description": "Contact",
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "last_update": {
                    "type": "string"
                },
                "password": {
                    "description": "Authentication",
                    "type": "string"
                },
                "role": {
                    "description": "Authorization",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Role"
                        }
                    ]
                },
                "totp_enabled": {
                    "type": "boolean"
                },
                "username": {
                    "description": "Primary key",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT to login",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}